release v1.2
* Mention lanelet2 in README.rst
* switch to 64bit node ids for supporting maps with large id values
* add all2all shortest_path method
* Guard against rogue lanelet ids and remove log spam
* Allow lanelets to relate to other lanelets
* Improve event regions
* Add new event region types overtaking and speedlimit
* Add .reverse() to Lanelet
* add traffic lights support
* add lanelet_checker tool

+++ src/liblanelet/LLTree.cpp
@@ -1,33 +1,30 @@
 // this is for emacs file handling -*- mode: c++; indent-tabs-mode: nil -*-
 
 // -- BEGIN LICENSE BLOCK ----------------------------------------------
-// Copyright (c) 2017, FZI Forschungszentrum Informatik
-// All rights reserved.
+// Copyright (c) 2018, FZI Forschungszentrum Informatik
 //
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
+// Redistribution and use in source and binary forms, with or without modification, are permitted
+// provided that the following conditions are met:
 //
-// * Redistributions of source code must retain the above copyright notice, this
-//   list of conditions and the following disclaimer.
+// 1. Redistributions of source code must retain the above copyright notice, this list of conditions
+//    and the following disclaimer.
 //
-// * Redistributions in binary form must reproduce the above copyright notice,
-//   this list of conditions and the following disclaimer in the documentation
-//   and/or other materials provided with the distribution.
+// 2. Redistributions in binary form must reproduce the above copyright notice, this list of
+//    conditions and the following disclaimer in the documentation and/or other materials provided
+//    with the distribution.
 //
-// * Neither the name of the copyright holder nor the names of its
-//   contributors may be used to endorse or promote products derived from
-//   this software without specific prior written permission.
+// 3. Neither the name of the copyright holder nor the names of its contributors may be used to
+//    endorse or promote products derived from this software without specific prior written
+//    permission.
 //
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-// OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
+// IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
+// FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+// WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
+// WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 // -- END LICENSE BLOCK ------------------------------------------------
 
 //----------------------------------------------------------------------
@@ -48,7 +45,7 @@ using namespace LLet;
 
 void LLTree::insert(const lanelet_ptr_t& obj)
 {
-    int32_t index = _lanelets.size();
+    int64_t index = _lanelets.size();
     _lanelets.push_back( obj );
     const BoundingBox bb = obj->bb();
     double min_data[2] = {bb.get<BoundingBox::SOUTH>(), bb.get<BoundingBox::WEST>()};
@@ -59,9 +56,9 @@ void LLTree::insert(const lanelet_ptr_t& obj)
 namespace
 {
 
-    bool callback(int32_t index, void* data )
+    bool callback(int64_t index, void* data )
     {
-        std::vector< int32_t >* indices = static_cast< std::vector< int32_t >* >(data);
+        std::vector< int64_t >* indices = static_cast< std::vector< int64_t >* >(data);
         indices->push_back(index);
         return true;
     }
@@ -73,14 +70,14 @@ std::vector< lanelet_ptr_t > LLTree::query( const BoundingBox& bb) const
     double min_data[2] = {bb.get<BoundingBox::SOUTH>(), bb.get<BoundingBox::WEST>()};
     double max_data[2] = {bb.get<BoundingBox::NORTH>(), bb.get<BoundingBox::EAST>()};
 
-    std::vector< int32_t > indices;
+    std::vector< int64_t > indices;
     _tree.Search(min_data, max_data, callback, &indices);
 
     std::vector< lanelet_ptr_t > result(indices.size());
 
-    // this replaces: std::transform(indices.begin(), indices.end(), result.begin(), [this](int32_t index){ return _lanelets[index];});
+    // this replaces: std::transform(indices.begin(), indices.end(), result.begin(), [this](int64_t index){ return _lanelets[index];});
     size_t i = 0;
-    BOOST_FOREACH(int32_t index, indices)
+    BOOST_FOREACH(int64_t index, indices)
     {
       result[i++] = _lanelets[index];
     }
